backend/upload-chapter/requirements.txt
psycopg2-binary==2.9.9
Pillow==10.4.0
